requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
gitpython>=3.1.40
faiss-cpu>=1.7.4
tiktoken>=0.5.1
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
rich>=13.5.0
typer>=0.9.0
pytest>=7.4.0
cydifflib>=0.1.2
pygments>=2.16.1
